(fix) adjust offset datatype to avoid int overflows (fixes #3628)

+++ SoObjects/Appointments/SOGoAppointmentFolder.m
@@ -766,7 +766,7 @@ static Class iCalEventK = nil;
   NSCalendarDate *date;
   NSNumber *dateValue;
   BOOL isAllDay;
-  unsigned int offset;
+  NSInteger offset;
 
   isAllDay = [[theRecord objectForKey: @"c_isallday"] boolValue];
   record = [[theRecord mutableCopy] autorelease];
